requirements.txt
kaggle==1.5.13
matplotlib==3.8.0
sympy==1.12
mlflow==2.9.2
wandb==0.16.1
sqlalchemy==2.0.23
prometheus_client==0.19.0
python-multipart==0.0.6
Pillow==10.1.0
torch
torchvision
fastapi
uvicorn[standard]==0.24.0.post1
